requirements.txt
# requirements.txt
torch>=1.8.0
tqdm>=4.62.0
biopython>=1.79
ml_collections>=0.1.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0